Make sortables generic
