+++ saleor/static/dashboard-next/components/SortableTable/index.ts
@@ -1,0 +1,5 @@
+export * from "./SortableTableBody";
+export { default as SortableTableBody } from "./SortableTableBody";
+
+export * from "./SortableTableRow";
+export { default as SortableTableRow } from "./SortableTableRow";
